+++ src/Controller/DashboardsController.php
@@ -1,0 +1,116 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Controller;
+
+use App\Model\Entity\Dashboard;
+use Cake\Http\Exception\ForbiddenException;
+
+/**
+ * Dashboards Controller
+ *
+ * @property \App\Model\Table\DashboardsTable $Dashboards
+ */
+class DashboardsController extends AppController
+{
+    /**
+     * Index method
+     *
+     * @return \Cake\Http\Response|null|void Renders view
+     */
+    public function index()
+    {
+        $user = $this->request
+            ->getAttribute('identity');
+        $query = $this->Dashboards->find()->where(['user_id' => $user->id]);
+        $dashboards = $this->paginate($query);
+
+        $this->set(compact('dashboards'));
+    }
+
+    /**
+     * View method
+     *
+     * @param string|null $id Dashboard id.
+     * @return \Cake\Http\Response|null|void Renders view
+     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
+     */
+    public function view($id = null)
+    {
+        $dashboard = $this->Dashboards->get($id, contain: ['Users']);
+        $user = $this->request->getAttribute('identity');
+        if ($user->id !== $dashboard->user_id) {
+            throw new ForbiddenException(__('You are not allowed to access this dashboard.'));
+        }
+        $this->set(compact('dashboard'));
+    }
+
+    /**
+     * Add method
+     *
+     * @return \Cake\Http\Response|null|void Redirects on successful add, renders view otherwise.
+     */
+    public function add()
+    {
+        /** @var Dashboard $dashboard */
+        $dashboard = $this->Dashboards->newEmptyEntity();
+        if ($this->request->is('post')) {
+            $dashboard = $this->Dashboards->patchEntity($dashboard, $this->request->getData());
+            $dashboard->user_id = $this->request->getAttribute('identity')->getIdentifier();
+            if ($this->Dashboards->save($dashboard)) {
+                $this->Flash->success(__('The dashboard has been saved.'));
+
+                return $this->redirect(['action' => 'index']);
+            }
+            $this->Flash->error(__('The dashboard could not be saved. Please, try again.'));
+        }
+        $this->set(compact('dashboard'));
+    }
+
+    /**
+     * Edit method
+     *
+     * @param string|null $id Dashboard id.
+     * @return \Cake\Http\Response|null|void Redirects on successful edit, renders view otherwise.
+     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
+     */
+    public function edit($id = null)
+    {
+        $dashboard = $this->Dashboards->get($id, contain: []);
+        if ($this->request->is(['patch', 'post', 'put'])) {
+            $dashboard = $this->Dashboards->patchEntity($dashboard, $this->request->getData());
+            $dashboard->user_id = $this->request->getAttribute('identity')->getIdentifier();
+            if ($this->Dashboards->save($dashboard)) {
+                $this->Flash->success(__('The dashboard has been saved.'));
+
+                return $this->redirect(['action' => 'index']);
+            }
+            $this->Flash->error(__('The dashboard could not be saved. Please, try again.'));
+        }
+        $this->set(compact('dashboard'));
+    }
+
+    /**
+     * Delete method
+     *
+     * @param string|null $id Dashboard id.
+     * @return \Cake\Http\Response|null Redirects to index.
+     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
+     */
+    public function delete($id = null)
+    {
+        $this->request->allowMethod(['post', 'delete']);
+        $dashboard = $this->Dashboards->get($id);
+        $user = $this->request->getAttribute('identity');
+        if ($user->id !== $dashboard->user_id) {
+            throw new ForbiddenException(__('You are not allowed to delete this dashboard.'));
+        }
+        if ($this->Dashboards->delete($dashboard)) {
+            $this->Flash->success(__('The dashboard has been deleted.'));
+        } else {
+            $this->Flash->error(__('The dashboard could not be deleted. Please, try again.'));
+        }
+
+        return $this->redirect(['action' => 'index']);
+    }
+}