+++ CHANGELOG.md
@@ -13,15 +13,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add this Changelog #6
 - Implement authentication with registration, login and logout through cakedc/users plugin #4 #5
 - Add Start page with link to registration and login #3
+- Add Dashboard CRUD #11
 
 ### Changed
+
+- Change user ID fields to integer #11
 
 ### Deprecated
 
 ### Removed
 
--Remove GitHub related files from CakePHP default app template
+- Remove GitHub related files from CakePHP default app template
 
 ### Fixed
+
+- Correct user_id field name
 
 ### Security